# Swagger.yml
...
openapi: '3.0.3'
info:
  title: FrogCrew API
  description: This is the FrogCrew API
  termsOfService: http://swagger.io/terms/
  contact:
    name: Kate Bednarz
    email: kate.bednarz@tcu.edu
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local server
...
paths:
...
  /scheduledGames/drop:
    post:
      tags:
        - scheduledGames
      summary: Add shift to trade board
      description: Add a scheduled game to the trade board.
      requestBody:
        required: true
        description: relevant information for dropping a shift a game
        content:
          application/json:
            example:
              userId: 1
              gameId: 1
              positionId: 1
      responses:
        '200':
          description: Scheduled game dropped
          content:
            application/json:
              example:
                flag: true
                code: 200
                message: Drop Success
                data:
                  tradeId: 1
                  dropperId: 1
                  gameId: 1
                  position: DIRECTOR
                  status: AVAILABLE 
                  receiverId: null
        '404':
          description: user, game, or shift not found
          content:
            application/json:
              examples:
                UserNotFound:
                  value:
                    flag: false
                    code: 404
                    message: Could not find user with id 1
                    data: null
                GameNotFound:
                  value:
                    flag: false
                    code: 404
                    message: Could not find game with id 1
                    data: null
                ShiftNotFound:
                  value:
                    flag: false
                    code: 404
                    message: Could not find shift
                    data: null
  /scheduledGames/pickup/{tradeId}/{userId}:
    put:
      tags:
        - scheduledGames
      summary: Request to pickup a game
      description: Request to pickup a game from the trade board.
      parameters:
        - in: path
          name: tradeId
          description: id of the trade board entry
          required: true
          schema:
            type: string
            example: 1
        - in: path
          name: userId
          description: id of the user picking up the shift
          required: true
          schema:
            type: string
            example: 1
      responses:
        '200':
          description: Game pickup requested
          content:
            application/json:
              example:
                flag: true
                code: 200
                message: Request Success
                data:
                  tradeId: 1
                  dropperId: 1
                  gameId: 1
                  position: DIRECTOR
                  status: awaiting approval
                  receiverId: 2
        '400':
          description: bad input parameter
          content:
            application/json:
              example:
                flag: false
                code: 400
                message: User is not qualified for this position
                data: null
        '404':
          description: user or trade not found
          content:
            application/json:
              examples:
                UserNotFound:
                  value:
                    flag: false
                    code: 404
                    message: Could not find user with id 1
                    data: null
                TradeNotFound:
                  value:
                    flag: false
                    code: 404
                    message: Could not find trade with id 1
                    data: null
  /scheduledGames/approve/{tradeId}:
    put:
      tags:
        - scheduledGames
      summary: Approve shift swap
      description: Approve a request to swap shift on tradeboard.
      parameters:
        - in: path
          name: tradeId
          description: id of the trade board entry
          required: true
          schema:
            type: string
            example: 1
      responses:
        '200':
          description: Game pickup approved
          content:
            application/json:
              example:
                flag: true
                code: 200
                message: Approval Success
                data:
                  tradeId: 1
                  dropperId: 1
                  gameId: 1
                  position: DIRECTOR
                  status: approved
                  receiverId: 2
        '404':
          description: user not found
          content:
            application/json:
              example:
                flag: false
                code: 404
                message: Could not find trade with id 1
                data: null
...